+++ presentation/src/main/java/com/threegap/bitnagil/presentation/emotion/EmotionScreen.kt
@@ -15,7 +15,7 @@ import com.threegap.bitnagil.presentation.emotion.component.template.EmotionReco
 import com.threegap.bitnagil.presentation.emotion.component.template.SimpleEmotionSelectionScreen
 import com.threegap.bitnagil.presentation.emotion.component.template.SwipeEmotionSelectionScreen
 import com.threegap.bitnagil.presentation.emotion.model.EmotionScreenStep
-import com.threegap.bitnagil.presentation.emotion.model.mvi.EmotionSideEffect
+import com.threegap.bitnagil.presentation.emotion.contract.EmotionSideEffect
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
