+++ presentation/src/main/java/com/threegap/bitnagil/presentation/emotion/component/template/SimpleEmotionSelectionScreen.kt
@@ -25,7 +25,7 @@ import com.threegap.bitnagil.presentation.emotion.component.atom.EmotionMarbleIm
 import com.threegap.bitnagil.presentation.emotion.model.EmotionImageUiModel
 import com.threegap.bitnagil.presentation.emotion.model.EmotionScreenStep
 import com.threegap.bitnagil.presentation.emotion.model.EmotionUiModel
-import com.threegap.bitnagil.presentation.emotion.model.mvi.EmotionState
+import com.threegap.bitnagil.presentation.emotion.contract.EmotionState
 
 @Composable
 fun SimpleEmotionSelectionScreen(
