+++ presentation/src/main/java/com/threegap/bitnagil/presentation/emotion/contract/EmotionSideEffect.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.presentation.emotion.contract
+
+sealed interface EmotionSideEffect {
+    data object NavigateToBack : EmotionSideEffect
+    data class ShowToast(val message: String) : EmotionSideEffect
+}
