--- presentation/src/main/java/com/threegap/bitnagil/presentation/emotion/model/mvi/EmotionSideEffect.kt
@@ -1,6 +1,0 @@
-package com.threegap.bitnagil.presentation.emotion.model.mvi
-
-sealed class EmotionSideEffect {
-    data object NavigateToBack : EmotionSideEffect()
-    data class ShowToast(val message: String) : EmotionSideEffect()
-}
